backend/orders/requirements.txt
google-cloud-firestore==2.9.0
flask==2.2.2
gunicorn==20.1.0
